requirements.txt
google-genai
python-dotenv
click
pillow
pytest
pytest-mock
pytest-cov
